Some issues have been fixed with libgdbwrap. The section is correctly returned by the server.  However, we are still not able to break :(. By uncommenting the 2 lines in libelfsh/vectors_call.c we can see that the readmem hook changes at a certain point in time, returning fancy values and sefaulting right after.
[...]
dim[0]: 0, dim[1]: 2, fct: 0x804c576
Returning: 0x8198798
dim[0]: 0, dim[1]: 2, fct: 0x804c576
Returning: 0x8198798
dim[0]: 0, dim[1]: 2, fct: 0x804c576
Returning: 0x8198798
dim[0]: 0, dim[1]: 0, fct: 0xb7ec85c0
dim[0]: 0, dim[1]: 0, fct: 0xb7ec85c0
Segmentation fault (core dumped)


C0

+++ libgdbwrap/include/gdbwrapper.h
@@ -53,10 +53,10 @@ char             *gdbwrap_own_command(char *command, gdbwrap_t *desc);
 void             gdbwrap_test(gdbwrap_t *desc);
 gdbwrap_gdbreg32 *gdbwrap_readgenreg(gdbwrap_t *desc);
 void             gdbwrap_continue(gdbwrap_t *desc);
-char             *gdbwrap_readmemory(la32 linaddr, uint8_t bytes,
+char             *gdbwrap_readmemory(la32 linaddr, unsigned bytes,
 				     gdbwrap_t *desc);
 void             *gdbwrap_writememory(la32 linaddr, void *value,
-				      uint8_t bytes, gdbwrap_t *desc);
+				      unsigned bytes, gdbwrap_t *desc);
 void             gdbwrap_writereg(ureg32 regNum, la32 val, gdbwrap_t *desc);
 void             gdbwrap_signal(int signal, gdbwrap_t *desc);
 void             gdbwrap_stepi(gdbwrap_t *desc);
